\name{rdaTest-package}
\alias{rdaTest-package}
\docType{package}
\title{ Canonical redundancy analysis }
\description{
Function \code{\link{rdaTest}} computes a simple or partial canonical redundancy analysis.
Function \code{\link{plot.rdaTest}} produces triplots from the output list of \code{rdaTest}. }

\details{

\tabular{ll}{
Package: \tab rdaTest\cr
Type: \tab Package\cr
Version: \tab 1.6\cr
Date: \tab 2010-02-21\cr
License: \tab Copyleft\cr
LazyLoad: \tab yes\cr
}

Function \code{\link{rdaTest}} computes a simple or partial canonical redundancy analysis.
Function \code{\link{plot.rdaTest}} produces triplots from the output list of \code{\link{rdaTest}}.
}

\author{ Pierre Legendre and Sebastien Durand, Universite de Montreal }

\references{
Legendre, P. and Legendre, L. 1998. Numerical Ecology. 2nd English ed. Elsevier, Amsterdam. }

\keyword{ package }
