\name{rdaTest-package}
\alias{rdaTest-package}
\docType{package}
\title{ Canonical redundancy analysis }
\description{
Function \code{\link{rdaTest}} computes a simple or partial canonical redundancy analysis.
Function \code{\link{plot.rdaTest}} produces triplots from the output list of \code{rdaTest}.
Function \code{\link{test.axes}} tests the significance of the individual canonical axes in simple or partial RDA. }

\details{

\tabular{ll}{
Package: \tab rdaTest\cr
Type: \tab Package\cr
Version: \tab 1.7\cr
Date: \tab 2010-10-04\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}

Function \code{\link{rdaTest}} computes a simple or partial canonical redundancy analysis.
Function \code{\link{plot.rdaTest}} produces triplots from the output list of \code{\link{rdaTest}}.
Function \code{\link{test.axes}} tests the significance of the individual canonical axes in simple or partial RDA.
}

\author{ Pierre Legendre and Sebastien Durand, Universite de Montreal }

\references{
Legendre, P. and Legendre, L. 2012. Numerical Ecology. 3rd English ed. Elsevier, Amsterdam. }

\keyword{ package }
